+++ src/actions/utils/mod.rs
@@ -1,0 +1,2 @@
+/// Utilities for handling commands
+pub mod command;